requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
chromadb>=1.0.0
tinydb>=4.8.0
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
flair>=0.13.1
torch>=2.0.0
transformers>=4.30.0
seqeval>=1.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
